requirements.txt
pydantic>=2.0.0
pandas>=2.0.0
typer>=0.9.0
pytest>=7.4.0
pyyaml>=6.0.0
rapidfuzz>=3.0.0
requests>=2.31.0
streamlit>=1.31.0
numpy>=1.24.0

# Optional: For embedding-based semantic search
openai>=1.0.0

# Optional: For Anthropic LLM provider
anthropic>=0.18.0
